rename exampleTask.slice.ts to task.slice.ts
